% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eglv_measurements.R
\name{get_eglv_measurements}
\alias{get_eglv_measurements}
\title{Get water level, discharge or velocity measurements for selected EGLV gauges}
\usage{
get_eglv_measurements(x = NULL, par = "waterlevel")
}
\arguments{
\item{x}{Sf object containing gauges to be used for subsequent queries,
as provided by \code{get_eglv_gauges()}.}

\item{par}{character. Parameter to be queried: \code{"waterlevel"}, \code{"discharge"}, or \code{"velocity"}.}
}
\value{
List of xts objects.
}
\description{
Get water level, discharge or velocity measurements for selected EGLV gauges
}
\examples{
\dontrun{
gauge <- get_eglv_gauges() |> dplyr::filter(id == "10103")

get_eglv_measurements(gauge)
get_eglv_measurements(gauge, par = "discharge")

gauges <- get_eglv_gauges() |> dplyr::filter(waterbody == "Hammbach")

get_eglv_measurements(gauges)
}
}
\seealso{
\code{\link[=get_eglv_gauges]{get_eglv_gauges()}}
}
