% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eglv_meta.R
\name{get_eglv_meta}
\alias{get_eglv_meta}
\title{Get (extended) metadata for selected EGLV gauges}
\usage{
get_eglv_meta(x = NULL)
}
\arguments{
\item{x}{Sf object containing gauges to be used for subsequent queries,
as provided by \code{get_eglv_gauges()}.}
}
\value{
Sf object.
}
\description{
Get (extended) metadata for selected EGLV gauges
}
\examples{
\dontrun{
gauge <- get_eglv_gauges() |> dplyr::filter(id == "10103")
get_eglv_meta(gauge)

gauges <- get_eglv_gauges() |> dplyr::filter(waterbody == "Hammbach")
get_eglv_meta(gauges)
}
}
\seealso{
\code{\link[=get_eglv_gauges]{get_eglv_gauges()}}
}
